requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
tqdm>=4.60.0
tensorboard>=2.10.0
pyyaml>=6.0
